Debugged MATLAB bundle for rigid motion registration in the DRT domain.
Fixed a set of bugs, several of them related to the fact that MATLAB uses the x- and (-y)- axes for its indices. Also allowed non-standard angles sets in the DRTs, as well as the option to resample the projection slopes (with a set of interpolation methods) to allow for greater precision in rotation angle estimation.

**NOTE** Currently, the radonreg.m function defines several ad-hoc parameters as defaults, w/o an interactive way of setting them. This should be changed soon.